First version, yay!
Okay, for the current edit we have a loooooooooooooooooooot of things,
these include THE WHOLE FUDGEING PROJECT!

*ehum*

Well, if you want to set yourself into the code, I recommend you start
in index.php, and then move on to ImageManipulation.php and Upload.php.
They represent the basic functionality so far.

For design, read up on form.css and index.php

The only functionality avaliable right now is to upload images (by
post, an ajax option will be made avaliable, but it's not that much of
a priorioty) and manipulate them. The manipulator is just for making
thumbnails, I don't know if I will extend its use in the WORLD OF
TOMORROW.

Oh yeah, the SQL.php file is a simple framework to MySQLi, where
SQL::query and SQL::$session will be the only methods of use to you.

Have a day.