Tag dialog now autofocuses and has close/submit hotkeys.
When the tagging dialog is opened, the textbox now automatically receives the
keyboard focus. This required a slight, non-breaking change to the API for
modal dialogs; see monad() comment in js/main.js for more information.

Additionally, the tagging dialog now accepts two hotkeys; pressing escape is
equivalent to clicking the cancel button, and pressing return is equivalent to
pressing the submit button.