Use back button and link friendly AJAX for image viewing.
Viewing images now sets the URL hash value to the image's ID. A framework is
put in place to react to changes in the hash value. This means that you can
now link to individual images while still linking to the image board, and use
the back and forward buttons to navigate.

Unfortunately, the only cross browser way to respond to hash value changes is
by repeatedly polling the hash. This patch polls the hash every 500
milliseconds, which seems to be enough for reasonably smooth browsing while
still using very few CPU resources.